+++ src/data/models/StartListItem.kt
@@ -4,6 +4,6 @@ data class StartListItem(
         val id: String,
         val userId: String,
         val competitionId: String,
-        val startTime: Float,
+        val startTime: Long,
         val description: String
 )
